distance: unsigned -> double
git-svn-id: svn://scm.gforge.inria.fr/svnroot/paradiseo@2277 331e1502-861f-0410-8da2-ba01fb791d7f

+++ trunk/paradiseo-mo/tutorial/Lesson6/sampling.cpp
@@ -187,7 +187,7 @@ void main_function(int argc, char **argv)
     eoHammingDistance<Indi> distance; // Hamming distance
     Indi bestSolution(vecSize, true); // global optimum
 
-    moDistanceStat<Indi, unsigned> distStat(distance, bestSolution); // statistic
+    moDistanceStat<Indi, double> distStat(distance, bestSolution); // statistic
 
     // "statistic" of the solution
     moSolutionStat<Indi> solStat;
